requirements.txt
openai==1.102.0
google-genai==1.32.0